+++ Lab-02/Task2/src/StudentStack.java
@@ -2,10 +2,10 @@
 
 public class StudentStack {
     private String[] name=new String[50];
-    private int[] mark=new int[50];
+    private double[] mark=new double[50];
     private int count=0;
 
-    public void push(String name,int mark){
+    public void push(String name,double mark){
         if(count>=50){
             System.out.println("The stack is full");
             return;
